corr f_distribution_cdf params datatype (still not implemented in src yet )

+++ include/probability.h
@@ -51,6 +51,6 @@ double chi_square_cdf(double x, int k);
 
 double f_distribution_pdf(double x, double d1, double d2);
 
-double f_distribution_cdf(double x, int d1, int d2);
+double f_distribution_cdf(double x, double d1, double d2);
 
 #endif
